Add files via upload
I am very glad to share with you guys that I have completed my first task that was assigned to me by CodSoft.
It is a simple Number game where we have to draw random numbers from 0-100 and guess which number it is. 
So this task required me to write a program using Java.